+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ This document describes changes between each past release.
 **Bug fix**
 
 - Update the `last_modified` value when updating the collection status and signature (#97)
+- Prevents crash with events on ``default`` bucket on Kinto < 3.3
 - Trigger ``ResourceChanged`` events when the destination collection and records are updated
   during signing. This allows plugins like ``kinto-changes`` and ``kinto.plugins.history``
   to catch the changes (#101).